Add CollaboraOnlineWebViewKeyboardManager
That is a small iOS framework intended to make the keyboard management
saner in the JavaScript of Collabora Online when running embedded in
an iOS app. Such apps are the Collabora Office iOS app and the
Nextcloud app. A small test app that embeds a tiny html page is
included.

This is just a start, significant further improvements are planned.