Reduce API complexity and handle iframes
As this code is specifically for the JavaScript in Collabora Online
("loleaflet"), we don't need to take any JS function names as
parameters. We control Collabora Online and can hardcode it to use the
function names that we want.

Handle arbitrarily nested iframes for the communication in the iOS app
-> JS direction. Assume that only one document (either the top one or
some iframe) will have the window.COKbdMgrCallback function. The
complexity related to posting messages to iframes and listeing for
them is handled automatically.

The only thing an iOS app that embeds Collabora Online in a WKWebView
needs to do is to allocate a CollaboraOnlineWebViewKeyboardManager
object and initialise it for the WKWebView, as in:

    keyboardManager = [[CollaboraOnlineWebViewKeyboardManager alloc] initForWebView:webView];