Use UITextField instead of UIKeyInput for more (potential) featurefulness
Also, make the API from JavaScript more easily extendable.

The message passed to
window.webkit.messageHandlers.CollaboraOnlineWebViewKeyboardManager is
now a JavaScript object. The property 'command' should (for now) be
the string "display" or "hide". In the "display" case, the object can
also contain a 'type' property, which is one of the UIKeyboardType
typeds, without the "UIKeyboardType" prefix. (See
https://developer.apple.com/documentation/uikit/uikeyboardtype )