Subclass UITextView instead of UITextField for more featurefulness
Now we get predictive suggestions.

Extend the API so that one can feed a "seed" text and caret location
in it to CollaboraOnlineWebViewKeyboardManager that it can preted we
are editing. (It doesn't have to correspond to the actual text in
whatever context this is used, in loleaflet we can't know it at the
moment, for instance. But in the test.html document here we do know.)

Leave a handful of NSLog() debugging output for now.